DJANGO_SIS/djangoschool/requirements.txt
atomicwrites==1.4.0
crispy-bootstrap4==2025.6
diff-match-patch==20241021
django-allauth==65.9.0
django-bulk-signals==0.4.2
django-crispy-forms==2.5
django-daisy==1.0.26
django-debug-toolbar==6.0.0
django-extensions==4.1
django-formtools==2.5.1
django-htmx==1.23.2
django-import-export==4.3.7
django-simple-history==3.11.0
django-slick-reporting==1.3.1
django-widget-tweaks==1.5.0
pillow==12.0.0
pyls-spyder==0.4.0
pytz==2025.2
reportlab==4.4.7
setuptools==78.1.1
simplejson==3.20.2
tablib==3.8.0
wheel==0.45.1